# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Opora API
  version: 0.3.0
  description: Authenticated API for the Opora student document workspace.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Session created. Sets opaque HttpOnly session and CSRF cookies.
          content: { application/json: { schema: { $ref: "#/components/schemas/CurrentUser" } } }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        "204": { description: Session revoked and cookies cleared. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/auth/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [Auth]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AcceptInvitation" } } }
      responses:
        "200": { description: Invitation accepted and password set. }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
  /api/v1/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Active session and effective permissions.
          content: { application/json: { schema: { $ref: "#/components/schemas/CurrentUser" } } }
        "401": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateCurrentUser
      tags: [Auth]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateProfile" } } }
      responses:
        "200": { description: Own profile updated. }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/me/change-password:
    post:
      operationId: changeCurrentUserPassword
      tags: [Auth]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ChangePassword" } } }
      responses:
        "204": { description: Password changed and other sessions revoked. }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/students:
    get:
      operationId: listStudents
      tags: [Students]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Students visible to the current membership.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/Student" } } }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: createStudent
      tags: [Students]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateStudent" } } }
      responses:
        "201":
          description: Student created.
          content: { application/json: { schema: { $ref: "#/components/schemas/Student" } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/students/{studentId}:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: getStudent
      tags: [Students]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Tenant-scoped student.
          content: { application/json: { schema: { $ref: "#/components/schemas/Student" } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateStudent
      tags: [Students]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateStudent" } } }
      responses:
        "200": { description: Tenant-scoped student updated. }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/roles:
    get:
      operationId: listRoles
      tags: [Users]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Roles configured for the current organization.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/Role" } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/users:
    get:
      operationId: listUsers
      tags: [Users]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Users and active memberships in the current organization.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/OrganizationUser" } } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createUser
      tags: [Users]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateUser" } } }
      responses:
        "201":
          description: Invited account created. Delivery status reports whether SMTP accepted the message.
          content: { application/json: { schema: { $ref: "#/components/schemas/OrganizationUser" } } }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/users/{userId}/invitation:
    parameters: [{ $ref: "#/components/parameters/UserID" }]
    post:
      operationId: resendUserInvitation
      tags: [Users]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        "200": { description: Previous pending invitation revoked and a new email sent. }
        "403": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/users/{userId}:
    parameters: [{ $ref: "#/components/parameters/UserID" }]
    patch:
      operationId: updateUser
      tags: [Users]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateUser" } } }
      responses:
        "200":
          description: Tenant-scoped user profile and role updated.
          content: { application/json: { schema: { $ref: "#/components/schemas/OrganizationUser" } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/organization:
    patch:
      operationId: updateOrganization
      tags: [Organization]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties: { name: { type: string, minLength: 1, maxLength: 255 } }
      responses:
        "200": { description: Current tenant organization updated. }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/students/{studentId}/access:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentAccess
      tags: [Access]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Active additive grants for the student.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/StudentAccess" } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: setStudentAccess
      tags: [Access]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SetStudentAccess" } } }
      responses:
        "204": { description: Grants replaced atomically. }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateStudentAccess
      tags: [Access]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SetStudentAccess" } } }
      responses:
        "204": { description: Grants replaced atomically. }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/students/{studentId}/access/{userId}:
    parameters:
      - { $ref: "#/components/parameters/StudentID" }
      - { $ref: "#/components/parameters/UserID" }
    delete:
      operationId: revokeStudentAccess
      tags: [Access]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        "204": { description: All grants revoked. }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/audit:
    get:
      operationId: listAuditEvents
      tags: [Audit]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Append-only audit events for the current organization.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/AuditEvent" } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/students/{studentId}/documents:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentDocuments
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Authorized document workspace.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/Document" } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: uploadStudentDocument
      tags: [Documents]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary, description: DOCX or PDF, maximum 25 MB. }
                title: { type: string, maxLength: 255 }
                documentType: { type: string, maxLength: 100 }
                confidentialityLevel: { type: string, enum: [standard, restricted], default: standard }
      responses:
        "201":
          description: Scanned immutable version 1 stored in private S3.
          content: { application/json: { schema: { $ref: "#/components/schemas/Document" } } }
        "403": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/documents/{documentId}:
    parameters: [{ $ref: "#/components/parameters/DocumentID" }]
    get:
      operationId: getDocument
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Tenant-scoped document metadata.
          content: { application/json: { schema: { $ref: "#/components/schemas/Document" } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/documents/{documentId}/versions:
    parameters: [{ $ref: "#/components/parameters/DocumentID" }]
    get:
      operationId: listDocumentVersions
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Immutable version history, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/DocumentVersion" } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/documents/{documentId}/download:
    parameters: [{ $ref: "#/components/parameters/DocumentID" }]
    get:
      operationId: downloadCurrentDocumentVersion
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Authorized current version stream.
          content:
            application/pdf: { schema: { type: string, format: binary } }
            application/vnd.openxmlformats-officedocument.wordprocessingml.document: { schema: { type: string, format: binary } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/documents/{documentId}/preview:
    parameters: [{ $ref: "#/components/parameters/DocumentID" }]
    get:
      operationId: previewPDF
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Authorized inline PDF stream.
          content:
            application/pdf: { schema: { type: string, format: binary } }
        "403": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
  /api/v1/documents/{documentId}/versions/{versionId}/download:
    parameters: [{ $ref: "#/components/parameters/DocumentID" }, { $ref: "#/components/parameters/VersionID" }]
    get:
      operationId: downloadDocumentVersion
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Authorized immutable version stream. }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/documents/{documentId}/editor:
    parameters: [{ $ref: "#/components/parameters/DocumentID" }]
    get:
      operationId: getDocumentEditorConfig
      tags: [Documents]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: JWT-signed ONLYOFFICE config for the current DOCX version.
          content:
            application/json:
              schema:
                type: object
                required: [documentServerUrl, config]
                properties:
                  documentServerUrl: { type: string, format: uri }
                  config: { type: object, additionalProperties: true }
        "403": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
  /api/v1/dashboard:
    get:
      operationId: getDashboard
      tags: [Dashboard]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Tenant-scoped daily workspace, content: { application/json: { schema: { $ref: "#/components/schemas/Dashboard" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/students/{studentId}/notes:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentNotes
      tags: [Casework]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Visible student notes, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/StudentNote" } } } } } } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createStudentNote
      tags: [Casework]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateStudentNote" } } } }
      responses:
        "201": { description: Note created, content: { application/json: { schema: { $ref: "#/components/schemas/StudentNote" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/students/{studentId}/support-cases:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentSupportCases
      tags: [Casework]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Student support cases with goals and progress, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/SupportCase" } } } } } } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createStudentSupportCase
      tags: [Casework]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateSupportCase" } } } }
      responses:
        "201": { description: Support case created, content: { application/json: { schema: { $ref: "#/components/schemas/SupportCase" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/support-cases/{caseId}/goals:
    post:
      operationId: createSupportGoal
      tags: [Casework]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters: [{ name: caseId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateSupportGoal" } } } }
      responses:
        "201": { description: Goal created, content: { application/json: { schema: { $ref: "#/components/schemas/SupportGoal" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/support-cases/{caseId}:
    patch:
      operationId: updateSupportCaseStatus
      tags: [Casework]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters: [{ name: caseId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [status], properties: { status: { type: string, enum: [active, monitoring, completed] } } } } } }
      responses:
        "200": { description: Support case status updated, content: { application/json: { schema: { $ref: "#/components/schemas/SupportCase" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/support-goals/{goalId}:
    patch:
      operationId: updateSupportGoalStatus
      tags: [Casework]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters: [{ name: goalId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [status], properties: { status: { type: string, enum: [planned, in_progress, achieved, cancelled] } } } } } }
      responses:
        "200": { description: Support goal status updated, content: { application/json: { schema: { $ref: "#/components/schemas/SupportGoal" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/support-goals/{goalId}/progress:
    post:
      operationId: createGoalProgress
      tags: [Casework]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters: [{ name: goalId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateGoalProgress" } } } }
      responses:
        "201": { description: Progress entry created, content: { application/json: { schema: { $ref: "#/components/schemas/GoalProgress" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/meetings:
    get:
      operationId: listMeetings
      tags: [Meetings]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Meetings for visible students, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CouncilMeeting" } } } } } } }
  /api/v1/students/{studentId}/meetings:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentMeetings
      tags: [Meetings]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Student meetings, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CouncilMeeting" } } } } } } }
    post:
      operationId: createStudentMeeting
      tags: [Meetings]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateCouncilMeeting" } } } }
      responses:
        "201": { description: Meeting created, content: { application/json: { schema: { $ref: "#/components/schemas/CouncilMeeting" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/meetings/{meetingId}:
    parameters: [{ name: meetingId, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      operationId: getMeeting
      tags: [Meetings]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Tenant-scoped meeting, content: { application/json: { schema: { $ref: "#/components/schemas/CouncilMeeting" } } } }
        "403": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateMeeting
      tags: [Meetings]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateCouncilMeeting" } } } }
      responses:
        "200": { description: Meeting updated or completed, content: { application/json: { schema: { $ref: "#/components/schemas/CouncilMeeting" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/tasks:
    get:
      operationId: listMyTasks
      tags: [Tasks]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Tasks assigned to current user, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/StudentTask" } } } } } } }
  /api/v1/students/{studentId}/tasks:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentTasks
      tags: [Tasks]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Student tasks, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/StudentTask" } } } } } } }
    post:
      operationId: createStudentTask
      tags: [Tasks]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateStudentTask" } } } }
      responses:
        "201": { description: Task created, content: { application/json: { schema: { $ref: "#/components/schemas/StudentTask" } } } }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: opora_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    StudentID: { name: studentId, in: path, required: true, schema: { type: string, format: uuid } }
    DocumentID: { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
...
    UserID: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
...
    Error:
      description: Stable public error envelope.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties: { email: { type: string, format: email }, password: { type: string, format: password } }
    AcceptInvitation:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token: { type: string, minLength: 32 }
        password: { type: string, format: password, minLength: 8, description: Must contain at least one letter and one digit. }
...
    CurrentUser:
      type: object
      required: [id, email, displayName, lastName, firstName, organization, permissions]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        lastName: { type: string }
        firstName: { type: string }
        middleName: { type: [string, "null"] }
        organization:
          type: object
          required: [id, name]
          properties: { id: { type: string, format: uuid }, name: { type: string } }
        permissions: { type: array, items: { type: string } }
...